backend/requirements.txt
flask
requests
beautifulsoup4
sentence-transformers
faiss-cpu
groq
python-dotenv
tqdm
urllib3
